refactor(artifacts): share study text rendering
